refactor: relocate dedicated files of `Hero`
